+++ src/components/donation/Steps/Submit/Crypto/TxSubmit/index.ts
@@ -1,1 +1,1 @@
-export { default } from "./Breakdown";
+export { default } from "./TxSubmit";
